+++ telegram/passport/credentials.py
@@ -23,7 +23,7 @@ except ImportError:
     import json  # type: ignore[no-redef]
 
 from base64 import b64decode
-from typing import TYPE_CHECKING, Any, List, Optional, Tuple, Union, no_type_check
+from typing import TYPE_CHECKING, Any, List, Optional, no_type_check
 
 try:
     from cryptography.hazmat.backends import default_backend
@@ -41,24 +41,11 @@ except ImportError:
 
     CRYPTO_INSTALLED = False
 
-from telegram import TelegramError, TelegramObject
+from telegram import TelegramObject, PassportDecryptionError
 from telegram.utils.types import JSONDict
 
 if TYPE_CHECKING:
     from telegram import Bot
-
-
-class TelegramDecryptionError(TelegramError):
-    """Something went wrong with decryption."""
-
-    __slots__ = ('_msg',)
-
-    def __init__(self, message: Union[str, Exception]):
-        super().__init__(f"TelegramDecryptionError: {message}")
-        self._msg = str(message)
-
-    def __reduce__(self) -> Tuple[type, Tuple[str]]:
-        return self.__class__, (self._msg,)
 
 
 @no_type_check
@@ -77,7 +64,7 @@ def decrypt(secret, hash, data):
             b64decode it.
 
     Raises:
-        :class:`TelegramDecryptionError`: Given hash does not match hash of decrypted data.
+        :class:`PassportDecryptionError`: Given hash does not match hash of decrypted data.
 
     Returns:
         :obj:`bytes`: The decrypted data as bytes.
@@ -105,7 +92,7 @@ def decrypt(secret, hash, data):
     # If the newly calculated hash did not match the one telegram gave us
     if data_hash != hash:
         # Raise a error that is caught inside telegram.PassportData and transformed into a warning
-        raise TelegramDecryptionError(f"Hashes are not equal! {data_hash} != {hash}")
+        raise PassportDecryptionError(f"Hashes are not equal! {data_hash} != {hash}")
     # Return data without padding
     return data[data[0] :]
 
@@ -173,7 +160,7 @@ class EncryptedCredentials(TelegramObject):
         :obj:`str`: Lazily decrypt and return secret.
 
         Raises:
-            telegram.TelegramDecryptionError: Decryption failed. Usually due to bad
+            telegram.PassportDecryptionError: Decryption failed. Usually due to bad
                 private/public key but can also suggest malformed/tampered data.
         """
         if self._decrypted_secret is None:
@@ -195,7 +182,7 @@ class EncryptedCredentials(TelegramObject):
                 )
             except ValueError as exception:
                 # If decryption fails raise exception
-                raise TelegramDecryptionError(exception) from exception
+                raise PassportDecryptionError(exception) from exception
         return self._decrypted_secret
 
     @property
@@ -206,7 +193,7 @@ class EncryptedCredentials(TelegramObject):
             `decrypted_data.nonce`.
 
         Raises:
-            telegram.TelegramDecryptionError: Decryption failed. Usually due to bad
+            telegram.PassportDecryptionError: Decryption failed. Usually due to bad
                 private/public key but can also suggest malformed/tampered data.
         """
         if self._decrypted_data is None:
